+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Add Loki s3 bucket and permissions.
+
 ## [14.10.0] - 2023-01-31
 
 ### Changed